AnnotTextMarkup::draw: simplify unnecessary opacity codepath split
The difference between opacity = 1 and opacity < 1 were
that in the former case the multiply is applied, and
in the latter, it is wrapped a second time in a
createForm.

The wrapping seems to be a no-op, while not applying the
multiply flag is a bug. Hence, we can simplify this by
removing the if statement and aligning both paths.

Thank You, @lcrippa, for reporting this!

Closes: https://gitlab.freedesktop.org/poppler/poppler/-/issues/1619

Reported-by: lorenzo crippa <crippa.lorenzo@gmail.com>
Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>